Optimized collision and rendering
The play area is split horizontally and vertically resulting in four
sub-areas. A bacter can belong to 1, 2 or 4 sub-areas depending on its
position. Collision testing is done only between bacters that are in the
same sub-area.
Only bacters inside the play area are rendered.
The title of the game window has "Debug" in it when debug build.